appearance mod structure: collapse settings and appearance files
