+++ src/ui/appearance/mod.rs
@@ -1,3 +1,2 @@
 pub mod appearance;
 pub mod appearance_background;
-pub mod settings;
